incorporando middlewares y entendiendo como funcionan

+++ store/routes/products-router.js
@@ -13,9 +13,13 @@ const service = new ProductsService();
 // Products
 const rutaProducts= '/';
 
-const CallbackProducts = async (request, response)=>{
-const products = await service.find();
-  response.json(products)
+const CallbackProducts = async (request, response, next)=>{
+  try {
+    const products = await service.find();
+    response.json(products)
+  } catch (error) {
+    next(error)
+  }
 }
 router.get(rutaProducts, CallbackProducts) /* Se cambia la palabra app por router */
 
@@ -33,16 +37,20 @@ router.get(productFilter,CallbackFilter)
 // Recibiendo Parámetros
 
 const productId = '/:id' /* Los 2 puntos significa que es un parámetro*/
-const CallbackProductId = async (request, response) => {
-  const {id} = request.params;
-  const product = await service.findOne(id)
-  response.json(product)
+const CallbackProductId = async (request, response, next) => {
+  try {
+    const {id} = request.params;
+    const product = await service.findOne(id)
+    response.json(product)
+  } catch (error) {
+    next(error)
+  }
 }
 router.get( productId, CallbackProductId)
 
 /* POST */
 const routePostProduct = '/'
-const CallbackPostProduct = async (request, response)=>{
+const CallbackPostProduct = async (request, response, next)=>{
   try{
     const body = request.body
     const newProduct = await service.create(body)/* Body es la información que se le manda al servidor*/
@@ -51,6 +59,7 @@ const CallbackPostProduct = async (request, response)=>{
     response.status(400).json({
       message : error.message
     })
+    next(error)
   }
 }
 
@@ -59,11 +68,15 @@ router.post(routePostProduct, CallbackPostProduct)
 /* PUT */
 
 const routePutProduct = '/:id'
-const CallbackPutProduct = async (request, response)=>{
-  const { id } = request.params;
-  const body = request.body
-  const updateProduct = await service.update(id, body)
-  response.json(updateProduct)
+const CallbackPutProduct = async (request, response, next)=>{
+  try {
+    const { id } = request.params;
+    const body = request.body
+    const updateProduct = await service.update(id, body)
+    response.json(updateProduct)
+  } catch (error) {
+    next(error)
+  }
 }
 
 router.put(routePutProduct, CallbackPutProduct)
@@ -71,7 +84,7 @@ router.put(routePutProduct, CallbackPutProduct)
 /* PATCH */
 
 const routePatchProduct = '/:id'
-const  CallbackPatchProduct = async (request, response)=>{
+const  CallbackPatchProduct = async (request, response, next)=>{
   try{ /* Para capturar errores */
     const { id } = request.params;
   const body = request.body
@@ -81,6 +94,7 @@ const  CallbackPatchProduct = async (request, response)=>{
     response.status(404).json({
       message: err.message
     })
+    next(err);
   }
 
 }
@@ -89,16 +103,21 @@ router.patch(routePatchProduct, CallbackPatchProduct)
 
 /* Delete */
 
-const routeDeleteProduct = '/:id' /*Ruta exige el id*/
+const CallbackDeleteProduct = async (request, response, next) => {
+  try {
+    const { id } = request.params;
+    const product = await service.findOne(id);
 
-const CallbackDeleteProduct = async (request, response)=>{
-  const { id } = request.params;/*trae el id de los parámetros*/
-  const deleteProduct = await service.delete(id)
-  response.json(deleteProduct)
-}
-
-router.delete(routeDeleteProduct, CallbackDeleteProduct) /*Ejecuta el método Delete*/
-
-
+    if (!product) {
+      return response.status(404).json({
+        message: `Product with id ${id} not found`,
+      });
+    }
+    const deleteProduct = await service.delete(id);
+    response.json(deleteProduct);
+  } catch (error) {
+    next(error);
+  }
+};
 
 module.exports = router /* Se exporta el modulo completo */
